Move withSelect() inside table components
